Dedupe code: DBus and Lua based items
Second step towards code deduplication (Issue #13): this collects part of the code that can be shared among `dbus_condition` and `and dbus_event` and places it into a module in `common` named `dbusitem`; also consolidates the enums that are used in Lua based items into a module in `common` named `luaitems`.
