+++ src/common.rs
@@ -1120,4 +1120,531 @@ pub mod cmditem {
 
 
 
+#[allow(dead_code)]
+/// This module provides utilities for Lua based items
+pub mod luaitem {
+
+    /// The possible values to be checked from Lua
+    pub enum LuaValue {
+        LuaString(String),
+        LuaNumber(f64),
+        LuaBoolean(bool),
+    }
+
+    /// In case of failure, the reason will be one of the provided values
+    #[derive(Debug, PartialEq)]
+    pub enum FailureReason {
+        NoCheck,
+        NoFailure,
+        VariableMatch,
+        ScriptError,
+    }
+}
+
+
+
+#[allow(dead_code)]
+pub mod dbusitem {
+    use regex::Regex;
+    use std::collections::HashMap;
+    use std::hash::{Hash, Hasher};
+    use cfgmap::CfgValue;
+    use zbus;
+    use zbus::zvariant;
+    use crate::constants::*;
+
+    // this helper is just to avoid ugly code since the implementation of
+    // zbus used here still does not implement the `key_signature` method
+    // for dictionaries
+    pub fn key_signature(d: &zvariant::Dict) -> String {
+        String::from(d.signature().as_str().chars().nth(2).unwrap())
+    }
+
+
+    /// an enum to store the operators for checking signal parameters
+    #[derive(PartialEq, Hash, Clone)]
+    pub enum ParamCheckOperator {
+        Equal,              // "eq"
+        NotEqual,           // "neq"
+        Greater,            // "gt"
+        GreaterEqual,       // "ge"
+        Less,               // "lt"
+        LessEqual,          // "le"
+        Match,              // "match"
+        Contains,           // "contains"
+        NotContains,        // "ncontains"
+    }
+
+    /// an enum containing the value that the parameter should be checked 
+    /// against
+    pub enum ParameterCheckValue {
+        Boolean(bool),
+        Integer(i64),
+        Float(f64),
+        String(String),
+        Regex(Regex),
+    }
+
+    /// an enum containing the possible types of indexes for parameters
+    #[derive(Hash)]
+    pub enum ParameterIndex {
+        Integer(u64),
+        String(String),
+    }
+
+    /// a struct containing a single test to be performed against a signal
+    /// payload
+    ///
+    /// short explaination, so that I remember how to use it:
+    /// - `Index`: contains a list of indexes which specify, also for nested
+    ///            structures. This means that for an array of mappings it
+    ///            might be of the form `{ 1, 3, "somepos" }` where the first
+    ///            `1` is the argument index, the `3` is the array index, and
+    ///            `"somepos"` is the mapping index.
+    /// - `Operator`: the operator to check the payload against
+    /// - `Value`: the value to compare the parameter entry to
+    pub struct ParameterCheckTest {
+        pub index: Vec<ParameterIndex>,
+        pub operator: ParamCheckOperator,
+        pub value: ParameterCheckValue,
+    }
+
+    // implement the hash protocol for ParameterCheckTest
+    impl Hash for ParameterCheckTest {
+        fn hash<H: Hasher>(&self, state: &mut H) {
+            self.index.hash(state);
+            self.operator.hash(state);
+            match &self.value {
+                ParameterCheckValue::Boolean(x) => x.hash(state),
+                ParameterCheckValue::Integer(x) => x.hash(state),
+                ParameterCheckValue::Float(x) => x.to_bits().hash(state),
+                ParameterCheckValue::String(x) => x.hash(state),
+                ParameterCheckValue::Regex(x) => x.as_str().hash(state),
+            }
+        }
+    }
+
+    // allow a test to be easily cloned
+    impl Clone for ParameterCheckTest {
+        fn clone(&self) -> Self {
+            let mut index: Vec<ParameterIndex> = Vec::new();
+            for i in self.index.iter() {
+                index.push({
+                    match i {
+                        ParameterIndex::Integer(u) => ParameterIndex::Integer(*u),
+                        ParameterIndex::String(s) => ParameterIndex::String(s.clone()),
+                    }
+                });
+            }
+            let value = match &self.value {
+                ParameterCheckValue::Boolean(x) => ParameterCheckValue::Boolean(*x),
+                ParameterCheckValue::Integer(x) => ParameterCheckValue::Integer(*x),
+                ParameterCheckValue::Float(x) => ParameterCheckValue::Float(*x),
+                ParameterCheckValue::String(s) => ParameterCheckValue::String(s.clone()),
+                ParameterCheckValue::Regex(e) => ParameterCheckValue::Regex(e.clone()),
+            };
+
+            ParameterCheckTest {
+                index,
+                operator: self.operator.clone(),
+                value,
+            }
+        }
+    }
+
+    /// a trait that defines containable types: implementations are provided for
+    /// all types found in the `ParameterCheckValue` enum defined above
+    pub trait Containable {
+        fn is_contained_in(&self, v: &zvariant::Value) -> bool;
+    }
+
+    // implementations: dictionary value lookup will be provided as soon as there
+    // will be a way, in _zbus_, to at least retrieve the dictionary keys (if not
+    // directly the mapped values) in order to compare the contents with the value
+    impl Containable for bool {
+        fn is_contained_in(&self, v: &zvariant::Value) -> bool {
+            match v {
+                zvariant::Value::Array(a) => {
+                    a.contains(&zvariant::Value::from(self))
+                }
+                _ => false
+            }
+        }
+    }
+
+    impl Containable for i64 {
+        fn is_contained_in(&self, v: &zvariant::Value) -> bool {
+            match v {
+                zvariant::Value::Array(a) => {
+                    // to handle this we transform the array into a new array of
+                    // i64 that is created to test for inclusion, and large u64
+                    // numbers are be automatically discarded and set to `None`
+                    // which is never matched
+                    let testv: Vec<Option<i64>>;
+                    match a.element_signature().as_str() {
+                        "y" => {    // BYTE
+                            testv = a.iter().map(|x| {
+                                if let zvariant::Value::U8(z) = x {
+                                    Some(i64::from(*z))
+                                } else {
+                                    None
+                                }
+                            }).collect();
+                        }
+                        "n" => {    // INT16
+                            testv = a.iter().map(|x| {
+                                if let zvariant::Value::I16(z) = x {
+                                    Some(i64::from(*z))
+                                } else {
+                                    None
+                                }
+                            }).collect();
+                        }
+                        "q" => {    // UINT16
+                            testv = a.iter().map(|x| {
+                                if let zvariant::Value::I16(z) = x {
+                                    Some(i64::from(*z))
+                                } else {
+                                    None
+                                }
+                            }).collect();
+                        }
+                        "i" => {    // INT32
+                            testv = a.iter().map(|x| {
+                                if let zvariant::Value::I32(z) = x {
+                                    Some(i64::from(*z))
+                                } else {
+                                    None
+                                }
+                            }).collect();
+                        }
+                        "u" => {    // UINT32
+                            testv = a.iter().map(|x| {
+                                if let zvariant::Value::U32(z) = x {
+                                    Some(i64::from(*z))
+                                } else {
+                                    None
+                                }
+                            }).collect();
+                        }
+                        "x" => {    // INT64
+                            testv = a.iter().map(|x| {
+                                if let zvariant::Value::I64(z) = x {
+                                    Some(i64::from(*z))
+                                } else {
+                                    None
+                                }
+                            }).collect();
+                        }
+                        "t" => {    // UINT64
+                            // this is the tricky one, but since we know that big
+                            // unsigned integer surely do not match the provided
+                            // value, we just convert them to `None` here, which
+                            // will never match
+                            testv = a.iter().map(|x| {
+                                if let zvariant::Value::U64(z) = x {
+                                    if *z <= i64::MAX as u64 {
+                                        Some(*z as i64)
+                                    } else {
+                                        None
+                                    }
+                                } else {
+                                    None
+                                }
+                            }).collect();
+                        }
+                        _ => { return false; }
+                    }
+                    testv.contains(&Some(*self))
+                }
+                _ => false
+            }
+        }
+    }
+
+    impl Containable for f64 {
+        fn is_contained_in(&self, v: &zvariant::Value) -> bool {
+            match v {
+                zvariant::Value::Array(a) => {
+                    a.contains(&zvariant::Value::from(*self))
+                }
+                _ => false
+            }
+        }
+    }
+
+    // String is a particular case, because it has to look for presence in arrays
+    // (both of `Str` and `ObjectPath`) or, alternatively, to match a substring
+    // of the returned `Str` or `ObjectPath`
+    impl Containable for String {
+        fn is_contained_in(&self, v: &zvariant::Value) -> bool {
+            match v {
+                zvariant::Value::Str(s) => {
+                    s.as_str().contains(self.as_str())
+                }
+                zvariant::Value::ObjectPath(s) => {
+                    s.as_str().contains(self.as_str())
+                }
+                zvariant::Value::Array(a) => {
+                    match a.element_signature().as_str() {
+                        "s" => {
+                            a.contains(&zvariant::Value::from(self))
+                        }
+                        "o" => {
+                            let o = zvariant::ObjectPath::try_from(self.as_str());
+                            if let Ok(o) = o {
+                                a.contains(&zvariant::Value::from(o))
+                            } else {
+                                false
+                            }
+                        }
+                        _ => false
+                    }
+                }
+                // currently used version of Dict (the one in zbus 3.x) does not
+                // allow to search the keys as set or list, thus the easiest test
+                // that can be made is retrieving a value and checking for errors
+                // and that the result is something
+                zvariant::Value::Dict(d) => {
+                    match key_signature(d).as_str() {
+                        "s" => {
+                            let k = zvariant::Str::from(self.as_str());
+                            let res: Result<Option<&zvariant::Value>, zvariant::Error> = d.get(&k);
+                            if let Ok(res) = res {
+                                if res.is_some() {
+                                    true
+                                } else {
+                                    false
+                                }
+                            } else {
+                                false
+                            }
+                        }
+                        "o" => {
+                            let k = zvariant::ObjectPath::try_from(self.as_str());
+                            if k.is_err() {
+                                return false
+                            }
+                            let res: Result<Option<&zvariant::Value>, zvariant::Error> = d.get(&k.unwrap());
+                            if let Ok(res) = res {
+                                if res.is_some() {
+                                    true
+                                } else {
+                                    false
+                                }
+                            } else {
+                                false
+                            }
+                        }
+                        _ => false
+                    }
+                }
+                _ => false
+            }
+        }
+    }
+
+    // the following is totally arbitrary and will actually not be used: it is
+    // provided here only in order to complete the "required" implementations
+    impl Containable for Regex {
+        fn is_contained_in(&self, v: &zvariant::Value) -> bool {
+            match v {
+                zvariant::Value::Array(a) => {
+                    for elem in a.to_vec() {
+                        if let zvariant::Value::Str(s) = elem {
+                            if self.is_match(s.as_str()) {
+                                return true;
+                            }
+                        }
+                    }
+                    false
+                }
+                _ => false
+            }
+        }
+    }
+
+    // the trait used to convert values to `zvariant::Value`
+    pub trait ToVariant {
+        fn to_variant(&self) -> Option<zvariant::Value>;
+    }
+
+    impl ToVariant for bool {
+        fn to_variant(&self) -> Option<zvariant::Value> {
+            Some(zvariant::Value::Bool(*self))
+        }
+    }
+
+    impl ToVariant for i64 {
+        fn to_variant(&self) -> Option<zvariant::Value> {
+            Some(zvariant::Value::I64(*self))
+        }
+    }
+
+    impl ToVariant for f64 {
+        fn to_variant(&self) -> Option<zvariant::Value> {
+            Some(zvariant::Value::F64(*self))
+        }
+    }
+
+    impl ToVariant for str {
+        fn to_variant(&self) -> Option<zvariant::Value> {
+            let s = &self.to_string();
+            if s.starts_with('\\') {
+                let rest = s.clone().split_off(2);
+                if s.starts_with("\\b") {
+                    let rest = rest.to_lowercase();
+                    if rest == "true" || rest == "1" {
+                        return Some(zvariant::Value::Bool(true));
+                    } else if rest == "false" || rest == "0" {
+                        return Some(zvariant::Value::Bool(false));
+                    } else {
+                        return None;
+                    }
+                } else if s.starts_with("\\y") {
+                    if let Ok(v) = rest.parse::<u8>() {
+                        return Some(zvariant::Value::U8(v));
+                    } else {
+                        return None;
+                    }
+                } else if s.starts_with("\\n") {
+                    if let Ok(v) = rest.parse::<i16>() {
+                        return Some(zvariant::Value::I16(v));
+                    } else {
+                        return None;
+                    }
+                } else if s.starts_with("\\q") {
+                    if let Ok(v) = rest.parse::<u16>() {
+                        return Some(zvariant::Value::U16(v));
+                    } else {
+                        return None;
+                    }
+                } else if s.starts_with("\\i") {
+                    if let Ok(v) = rest.parse::<i32>() {
+                        return Some(zvariant::Value::I32(v));
+                    } else {
+                        return None;
+                    }
+                } else if s.starts_with("\\u") {
+                    if let Ok(v) = rest.parse::<u32>() {
+                        return Some(zvariant::Value::U32(v));
+                    } else {
+                        return None;
+                    }
+                } else if s.starts_with("\\x") {
+                    if let Ok(v) = rest.parse::<i64>() {
+                        return Some(zvariant::Value::I64(v));
+                    } else {
+                        return None;
+                    }
+                } else if s.starts_with("\\t") {
+                    if let Ok(v) = rest.parse::<u64>() {
+                        return Some(zvariant::Value::U64(v));
+                    } else {
+                        return None;
+                    }
+                } else if s.starts_with("\\d") {
+                    if let Ok(v) = rest.parse::<f64>() {
+                        return Some(zvariant::Value::F64(v));
+                    } else {
+                        return None;
+                    }
+                } else if s.starts_with("\\s") {
+                    Some(zvariant::Value::new(rest.clone()))
+                } else if s.starts_with("\\o") {
+                    // here we check it, having the RE at hand
+                    if RE_DBUS_OBJECT_PATH.is_match(&rest) {
+                        Some(zvariant::Value::new(
+                            zvariant::ObjectPath::try_from(rest.clone()).unwrap()))
+                    } else {
+                        None
+                    }
+                } else if s.starts_with("\\\\") {
+                    Some(zvariant::Value::new(String::from("\\") + &rest))
+                } else {
+                    Some(zvariant::Value::new(String::from(s)))
+                }
+            } else {
+                Some(zvariant::Value::new(String::from(s)))
+            }
+        }
+    }
+
+    impl<T> ToVariant for Vec<T>
+    where T: ToVariant {
+        fn to_variant(&self) -> Option<zvariant::Value> {
+            let mut a: Vec<zvariant::Value> = Vec::new();
+            for item in self.iter() {
+                if let Some(v) = item.to_variant() {
+                    a.push(v)
+                } else {
+                    return None;
+                }
+            }
+            Some(zvariant::Value::new(a))
+        }
+    }
+
+    // we only support maps where the key is a string
+    impl<T> ToVariant for HashMap<String, T>
+    where T: ToVariant {
+        fn to_variant(&self) -> Option<zvariant::Value> {
+            let mut d: HashMap<String, zvariant::Value> = HashMap::new();
+            for (key, item) in self.iter() {
+                if let Some(v) = item.to_variant() {
+                    d.insert(key.clone(), v);
+                } else {
+                    return None;
+                }
+            }
+            Some(zvariant::Value::new(d))
+        }
+    }
+
+    // this is necessary for the following conversion
+    impl ToVariant for zvariant::Value<'_> {
+        fn to_variant(&self) -> Option<zvariant::Value> {
+            Some(self.clone())
+        }
+    }
+
+    // and finally we support CfgValue, which is similar to a variant
+    impl ToVariant for CfgValue {
+        fn to_variant(&self) -> Option<zvariant::Value> {
+            if self.is_bool() {
+                self.as_bool().unwrap().to_variant()
+            } else if self.is_int() {
+                self.as_int().unwrap().to_variant()
+            } else if self.is_float() {
+                self.as_float().unwrap().to_variant()
+            } else if self.is_str() {
+                self.as_str().unwrap().to_variant()
+            } else if self.is_list() {
+                self.as_list().unwrap().to_variant()
+            } else if self.is_map() {
+                let map = self.as_map().unwrap();
+                let mut h: HashMap<String, zvariant::Value> = HashMap::new();
+                for key in map.keys() {
+                    if let Some(value) = map.get(key) {
+                        if let Some(v) = value.to_variant() {
+                            h.insert(key.clone(), v);
+                        } else {
+                            return None;
+                        }
+                    } else {
+                        return None;
+                    }
+                }
+                Some(zvariant::Value::new(h))
+            } else {
+                None
+            }
+        }
+    }
+
+
+}
+
+
+
 // end.
