+++ src/event/dbus_event.rs
@@ -25,11 +25,11 @@ use zbus::zvariant;
 use std::str::FromStr;
 use serde_json::value::Value;
 
-
 use super::base::Event;
 use crate::condition::registry::ConditionRegistry;
 use crate::condition::bucket_cond::ExecutionBucket;
 use crate::common::logging::{log, LogType};
+use crate::common::dbusitem::*;
 use crate::{cfg_mandatory, constants::*};
 
 use crate::cfghelp::*;
@@ -37,319 +37,6 @@ use crate::cfghelp::*;
 
 // see the DBus specification
 const DBUS_MAX_NUMBER_OF_ARGUMENTS: i64 = 63;
-
-
-/// an enum to store the operators for checking signal parameters
-#[derive(PartialEq, Hash, Clone)]
-enum ParamCheckOperator {
-    Equal,              // "eq"
-    NotEqual,           // "neq"
-    Greater,            // "gt"
-    GreaterEqual,       // "ge"
-    Less,               // "lt"
-    LessEqual,          // "le"
-    Match,              // "match"
-    Contains,           // "contains"
-    NotContains,        // "ncontains"
-}
-
-/// an enum containing the value that the parameter should be checked against
-enum ParameterCheckValue {
-    Boolean(bool),
-    Integer(i64),
-    Float(f64),
-    String(String),
-    Regex(Regex),
-}
-
-/// an enum containing the possible types of indexes for parameters
-#[derive(Hash)]
-enum ParameterIndex {
-    Integer(u64),
-    String(String),
-}
-
-/// a struct containing a single test to be performed against a signal payload
-///
-/// short explaination, so that I remember how to use it:
-/// - `Index`: contains a list of indexes which specify, also for nested
-///            structures. This means that for an array of mappings it might
-///            be of the form `{ 1, 3, "somepos" }` where the first `1` is the
-///            argument index, the `3` is the array index, and `"somepos"` is
-///            the mapping index.
-/// - `Operator`: the operator to check the payload against
-/// - `Value`: the value to compare the parameter entry to
-struct ParameterCheckTest {
-    index: Vec<ParameterIndex>,
-    operator: ParamCheckOperator,
-    value: ParameterCheckValue,
-}
-
-
-// implement the hash protocol for ParameterCheckTest
-impl Hash for ParameterCheckTest {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.index.hash(state);
-        self.operator.hash(state);
-        match &self.value {
-            ParameterCheckValue::Boolean(x) => x.hash(state),
-            ParameterCheckValue::Integer(x) => x.hash(state),
-            ParameterCheckValue::Float(x) => x.to_bits().hash(state),
-            ParameterCheckValue::String(x) => x.hash(state),
-            ParameterCheckValue::Regex(x) => x.as_str().hash(state),
-        }
-    }
-}
-
-impl Clone for ParameterCheckTest {
-    fn clone(&self) -> Self {
-        let mut index: Vec<ParameterIndex> = Vec::new();
-        for i in self.index.iter() {
-            index.push({
-                match i {
-                    ParameterIndex::Integer(u) => ParameterIndex::Integer(*u),
-                    ParameterIndex::String(s) => ParameterIndex::String(s.clone()),
-                }
-            });
-        }
-        let value = match &self.value {
-            ParameterCheckValue::Boolean(x) => ParameterCheckValue::Boolean(*x),
-            ParameterCheckValue::Integer(x) => ParameterCheckValue::Integer(*x),
-            ParameterCheckValue::Float(x) => ParameterCheckValue::Float(*x),
-            ParameterCheckValue::String(s) => ParameterCheckValue::String(s.clone()),
-            ParameterCheckValue::Regex(e) => ParameterCheckValue::Regex(e.clone()),
-        };
-
-        ParameterCheckTest {
-            index,
-            operator: self.operator.clone(),
-            value,
-        }
-    }
-}
-
-
-
-// this helper is just to avoid ugly code since the implementation of
-// zbus used here still does not implement the `key_signature` method
-// for dictionaries
-fn key_signature(d: &zvariant::Dict) -> String {
-    String::from(d.signature().as_str().chars().nth(2).unwrap())
-}
-
-
-
-/// a trait that defines containable types: implementations are provided for
-/// all types found in the `ParameterCheckValue` enum defined above
-trait Containable {
-    fn is_contained_in(&self, v: &zvariant::Value) -> bool;
-}
-
-// implementations: dictionary value lookup will be provided as soon as there
-// will be a way, in _zbus_, to at least retrieve the dictionary keys (if not
-// directly the mapped values) in order to compare the contents with the value
-impl Containable for bool {
-    fn is_contained_in(&self, v: &zvariant::Value) -> bool {
-        match v {
-            zvariant::Value::Array(a) => {
-                a.contains(&zvariant::Value::from(self))
-            }
-            _ => false
-        }
-    }
-}
-
-impl Containable for i64 {
-    fn is_contained_in(&self, v: &zvariant::Value) -> bool {
-        match v {
-            zvariant::Value::Array(a) => {
-                // to handle this we transform the array into a new array of
-                // i64 that is created to test for inclusion, and large u64
-                // numbers are be automatically discarded and set to `None`
-                // which is never matched
-                let testv: Vec<Option<i64>>;
-                match a.element_signature().as_str() {
-                    "y" => {    // BYTE
-                        testv = a.iter().map(|x| {
-                            if let zvariant::Value::U8(z) = x {
-                                Some(i64::from(*z))
-                            } else {
-                                None
-                            }
-                        }).collect();
-                    }
-                    "n" => {    // INT16
-                        testv = a.iter().map(|x| {
-                            if let zvariant::Value::I16(z) = x {
-                                Some(i64::from(*z))
-                            } else {
-                                None
-                            }
-                        }).collect();
-                    }
-                    "q" => {    // UINT16
-                        testv = a.iter().map(|x| {
-                            if let zvariant::Value::I16(z) = x {
-                                Some(i64::from(*z))
-                            } else {
-                                None
-                            }
-                        }).collect();
-                    }
-                    "i" => {    // INT32
-                        testv = a.iter().map(|x| {
-                            if let zvariant::Value::I32(z) = x {
-                                Some(i64::from(*z))
-                            } else {
-                                None
-                            }
-                        }).collect();
-                    }
-                    "u" => {    // UINT32
-                        testv = a.iter().map(|x| {
-                            if let zvariant::Value::U32(z) = x {
-                                Some(i64::from(*z))
-                            } else {
-                                None
-                            }
-                        }).collect();
-                    }
-                    "x" => {    // INT64
-                        testv = a.iter().map(|x| {
-                            if let zvariant::Value::I64(z) = x {
-                                Some(i64::from(*z))
-                            } else {
-                                None
-                            }
-                        }).collect();
-                    }
-                    "t" => {    // UINT64
-                        // this is the tricky one, but since we know that big
-                        // unsigned integer surely do not match the provided
-                        // value, we just convert them to `None` here, which
-                        // will never match
-                        testv = a.iter().map(|x| {
-                            if let zvariant::Value::U64(z) = x {
-                                if *z <= i64::MAX as u64 {
-                                    Some(*z as i64)
-                                } else {
-                                    None
-                                }
-                            } else {
-                                None
-                            }
-                        }).collect();
-                    }
-                    _ => { return false; }
-                }
-                testv.contains(&Some(*self))
-            }
-            _ => false
-        }
-    }
-}
-
-impl Containable for f64 {
-    fn is_contained_in(&self, v: &zvariant::Value) -> bool {
-        match v {
-            zvariant::Value::Array(a) => {
-                a.contains(&zvariant::Value::from(*self))
-            }
-            _ => false
-        }
-    }
-}
-
-// String is a particular case, because it has to look for presence in arrays
-// (both of `Str` and `ObjectPath`) or, alternatively, to match a substring
-// of the returned `Str` or `ObjectPath`
-impl Containable for String {
-    fn is_contained_in(&self, v: &zvariant::Value) -> bool {
-        match v {
-            zvariant::Value::Str(s) => {
-                s.as_str().contains(self.as_str())
-            }
-            zvariant::Value::ObjectPath(s) => {
-                s.as_str().contains(self.as_str())
-            }
-            zvariant::Value::Array(a) => {
-                match a.element_signature().as_str() {
-                    "s" => {
-                        a.contains(&zvariant::Value::from(self))
-                    }
-                    "o" => {
-                        let o = zvariant::ObjectPath::try_from(self.as_str());
-                        if let Ok(o) = o {
-                            a.contains(&zvariant::Value::from(o))
-                        } else {
-                            false
-                        }
-                    }
-                    _ => false
-                }
-            }
-            // currently used version of Dict (the one in zbus 3.x) does not
-            // allow to search the keys as set or list, thus the easiest test
-            // that can be made is retrieving a value and checking for errors
-            // and that the result is something
-            zvariant::Value::Dict(d) => {
-                match key_signature(d).as_str() {
-                    "s" => {
-                        let k = zvariant::Str::from(self.as_str());
-                        let res: Result<Option<&zvariant::Value>, zvariant::Error> = d.get(&k);
-                        if let Ok(res) = res {
-                            if res.is_some() {
-                                true
-                            } else {
-                                false
-                            }
-                        } else {
-                            false
-                        }
-                    }
-                    "o" => {
-                        let k = zvariant::ObjectPath::try_from(self.as_str());
-                        if k.is_err() {
-                            return false
-                        }
-                        let res: Result<Option<&zvariant::Value>, zvariant::Error> = d.get(&k.unwrap());
-                        if let Ok(res) = res {
-                            if res.is_some() {
-                                true
-                            } else {
-                                false
-                            }
-                        } else {
-                            false
-                        }
-                    }
-                    _ => false
-                }
-            }
-            _ => false
-        }
-    }
-}
-
-// the following is totally arbitrary and will actually not be used: it is
-// provided here only in order to complete the "required" implementations
-impl Containable for Regex {
-    fn is_contained_in(&self, v: &zvariant::Value) -> bool {
-        match v {
-            zvariant::Value::Array(a) => {
-                for elem in a.to_vec() {
-                    if let zvariant::Value::Str(s) = elem {
-                        if self.is_match(s.as_str()) {
-                            return true;
-                        }
-                    }
-                }
-                false
-            }
-            _ => false
-        }
-    }
-}
